Updated scheduler entities and repositories to use int IDs

+++ src/main/java/name/joseland/mal/automation/scheduler/db/CronExpressionRepository.java
@@ -7,6 +7,6 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface CronExpressionRepository extends JpaRepository<CronExpression, Long> {
+public interface CronExpressionRepository extends JpaRepository<CronExpression, Integer> {
 
 }
